@@ -1,81 +1,92 @@
 // Copyright 2019 Project March.
+#include "march_hardware_interface/march_hardware_interface.h"
+#include "march_hardware_interface/power_net_on_off_command.h"
+
+#include <sstream>
+#include <string>
 
 #include <joint_limits_interface/joint_limits.h>
 #include <joint_limits_interface/joint_limits_interface.h>
 #include <joint_limits_interface/joint_limits_urdf.h>
-#include <sstream>
-#include <string>
-
-#include <march_hardware/MarchRobot.h>
-#include <march_hardware/Joint.h>
-#include <march_hardware/ActuationMode.h>
-
-#include <march_hardware_interface/power_net_on_off_command.h>
-#include <march_hardware_interface/march_hardware_interface.h>
-
 #include <urdf/model.h>
+
+#include <march_hardware/ActuationMode.h>
+#include <march_hardware/Joint.h>
 
 using hardware_interface::JointHandle;
 using hardware_interface::JointStateHandle;
 using hardware_interface::PositionJointInterface;
 using joint_limits_interface::EffortJointSoftLimitsHandle;
-using joint_limits_interface::EffortJointSoftLimitsInterface;
 using joint_limits_interface::JointLimits;
 using joint_limits_interface::PositionJointSoftLimitsHandle;
-using joint_limits_interface::PositionJointSoftLimitsInterface;
 using joint_limits_interface::SoftJointLimits;
 using march::Joint;
 
-MarchHardwareInterface::MarchHardwareInterface(ros::NodeHandle& nh, AllowedRobot robotName)
-  : nh_(nh), marchRobot(HardwareBuilder(robotName).createMarchRobot())
+MarchHardwareInterface::MarchHardwareInterface(march::MarchRobot robot)
+  : march_robot_(std::move(robot))
+  , has_power_distribution_board_(march_robot_.getPowerDistributionBoard()->getSlaveIndex() != -1)
 {
 }
 
-void MarchHardwareInterface::init()
+bool MarchHardwareInterface::init(ros::NodeHandle& nh, ros::NodeHandle& /* robot_hw_nh */)
 {
   // Initialize realtime publisher for the IMotionCube states
-  imc_state_pub_ = RtPublisherPtr(
-      new realtime_tools::RealtimePublisher<march_shared_resources::ImcErrorState>(this->nh_, "/march/imc_states/", 4));
+  this->imc_state_pub_ = std::make_unique<realtime_tools::RealtimePublisher<march_shared_resources::ImcErrorState>>(
+      nh, "/march/imc_states/", 4);
 
-  after_limit_joint_command_pub_ = RtPublisherAfterLimitJointCommandPtr(
-      new realtime_tools::RealtimePublisher<march_shared_resources::AfterLimitJointCommand>(
-          this->nh_, "/march/controller/after_limit_joint_command/", 4));
+  this->after_limit_joint_command_pub_ =
+      std::make_unique<realtime_tools::RealtimePublisher<march_shared_resources::AfterLimitJointCommand>>(
+          nh, "/march/controller/after_limit_joint_command/", 4);
 
   // Start ethercat cycle in the hardware
-  this->marchRobot.startEtherCAT();
+  this->march_robot_.startEtherCAT();
 
   urdf::Model model;
   if (!model.initParam("/robot_description"))
   {
     ROS_ERROR("Failed to read the urdf from the parameter server.");
-    throw std::runtime_error("Failed to read the urdf from the parameter server.");
+    return false;
   }
 
   // Get joint names from urdf
-  for (auto const& urdfJoint : model.joints_)
+  for (const auto& urdf_joint : model.joints_)
   {
-    if (urdfJoint.second->type != urdf::Joint::FIXED)
+    if (urdf_joint.second->type != urdf::Joint::FIXED)
     {
-      joint_names_.push_back(urdfJoint.first);
+      joint_names_.push_back(urdf_joint.first);
     }
   }
 
-  nh_.setParam("/march/joint_names", joint_names_);
+  nh.setParam("/march/joint_names", joint_names_);
 
   num_joints_ = joint_names_.size();
 
   // Resize vectors
   joint_position_.resize(num_joints_);
+  joint_position_command_.resize(num_joints_);
   joint_velocity_.resize(num_joints_);
+  joint_velocity_command_.resize(num_joints_);
   joint_effort_.resize(num_joints_);
+  joint_effort_command_.resize(num_joints_);
   joint_temperature_.resize(num_joints_);
   joint_temperature_variance_.resize(num_joints_);
-  joint_position_command_.resize(num_joints_);
-  joint_velocity_command_.resize(num_joints_);
-  joint_effort_command_.resize(num_joints_);
   soft_limits_.resize(num_joints_);
 
-  for (int i = 0; i < num_joints_; ++i)
+  after_limit_joint_command_pub_->msg_.name.resize(num_joints_);
+  after_limit_joint_command_pub_->msg_.position_command.resize(num_joints_);
+  after_limit_joint_command_pub_->msg_.effort_command.resize(num_joints_);
+
+  imc_state_pub_->msg_.joint_names.resize(num_joints_);
+  imc_state_pub_->msg_.status_word.resize(num_joints_);
+  imc_state_pub_->msg_.detailed_error.resize(num_joints_);
+  imc_state_pub_->msg_.motion_error.resize(num_joints_);
+  imc_state_pub_->msg_.state.resize(num_joints_);
+  imc_state_pub_->msg_.detailed_error_description.resize(num_joints_);
+  imc_state_pub_->msg_.motion_error_description.resize(num_joints_);
+  imc_state_pub_->msg_.motor_current.resize(num_joints_);
+  imc_state_pub_->msg_.motor_voltage.resize(num_joints_);
+
+  for (size_t i = 0; i < num_joints_; ++i)
   {
     SoftJointLimits soft_limits;
     getSoftJointLimits(model.getJoint(joint_names_[i]), soft_limits);
@@ -87,34 +98,33 @@ void MarchHardwareInterface::init()
   initiateIMC();
 
   // Create march_pdb_state interface
-  MarchPdbStateHandle marchPdbStateHandle("PDBhandle", &power_distribution_board_read_,
-                                          &master_shutdown_allowed_command, &enable_high_voltage_command,
-                                          &power_net_on_off_command_);
-  march_pdb_interface.registerHandle(marchPdbStateHandle);
+  MarchPdbStateHandle march_pdb_state_handle("PDBhandle", &power_distribution_board_read_,
+                                             &master_shutdown_allowed_command_, &enable_high_voltage_command_,
+                                             &power_net_on_off_command_);
+  march_pdb_interface_.registerHandle(march_pdb_state_handle);
 
-  registerInterface(&march_temperature_interface);
-  registerInterface(&march_pdb_interface);
+  registerInterface(&march_temperature_interface_);
+  registerInterface(&march_pdb_interface_);
   registerInterface(&joint_state_interface_);
   registerInterface(&position_joint_interface_);
   registerInterface(&effort_joint_interface_);
-  registerInterface(&positionJointSoftLimitsInterface);
-  registerInterface(&effortJointSoftLimitsInterface);
+  registerInterface(&position_joint_soft_limits_interface_);
+  registerInterface(&effort_joint_soft_limits_interface_);
 
-  hasPowerDistributionBoard = marchRobot.getPowerDistributionBoard()->getSlaveIndex() != -1;
-  if (hasPowerDistributionBoard)
+  if (has_power_distribution_board_)
   {
-    for (int i = 0; i < num_joints_; i++)
+    for (size_t i = 0; i < num_joints_; i++)
     {
-      int netNumber = marchRobot.getJoint(joint_names_[i]).getNetNumber();
-      if (netNumber == -1)
+      int net_number = march_robot_.getJoint(joint_names_[i]).getNetNumber();
+      if (net_number == -1)
       {
-        std::ostringstream errorStream;
-        errorStream << "Joint " << joint_names_[i].c_str() << " has no net number";
-        throw std::runtime_error(errorStream.str());
+        std::ostringstream error_stream;
+        error_stream << "Joint " << joint_names_[i].c_str() << " has no net number";
+        throw std::runtime_error(error_stream.str());
       }
-      while (!marchRobot.getPowerDistributionBoard()->getHighVoltage().getNetOperational(netNumber))
+      while (!march_robot_.getPowerDistributionBoard()->getHighVoltage().getNetOperational(net_number))
       {
-        marchRobot.getPowerDistributionBoard()->getHighVoltage().setNetOnOff(true, netNumber);
+        march_robot_.getPowerDistributionBoard()->getHighVoltage().setNetOnOff(true, net_number);
         usleep(100000);
         ROS_WARN("[%s] Waiting on high voltage", joint_names_[i].c_str());
       }
@@ -126,12 +136,12 @@ void MarchHardwareInterface::init()
   }
 
   // Initialize interfaces for each joint
-  for (int i = 0; i < num_joints_; ++i)
+  for (size_t i = 0; i < num_joints_; ++i)
   {
-    march::Joint joint = marchRobot.getJoint(joint_names_[i]);
+    march::Joint joint = march_robot_.getJoint(joint_names_[i]);
     // Create joint state interface
-    JointStateHandle jointStateHandle(joint.getName(), &joint_position_[i], &joint_velocity_[i], &joint_effort_[i]);
-    joint_state_interface_.registerHandle(jointStateHandle);
+    JointStateHandle joint_state_handle(joint.getName(), &joint_position_[i], &joint_velocity_[i], &joint_effort_[i]);
+    joint_state_interface_.registerHandle(joint_state_handle);
 
     // Retrieve joint (soft) limits from the urdf
     JointLimits limits;
@@ -140,47 +150,47 @@ void MarchHardwareInterface::init()
     if (joint.getActuationMode() == march::ActuationMode::position)
     {
       // Create position joint interface
-      JointHandle jointPositionHandle(jointStateHandle, &joint_position_command_[i]);
-      position_joint_interface_.registerHandle(jointPositionHandle);
+      JointHandle joint_position_handle(joint_state_handle, &joint_position_command_[i]);
+      position_joint_interface_.registerHandle(joint_position_handle);
 
       // Create joint limit interface
-      PositionJointSoftLimitsHandle jointLimitsHandle(jointPositionHandle, limits, soft_limits_[i]);
-      positionJointSoftLimitsInterface.registerHandle(jointLimitsHandle);
+      PositionJointSoftLimitsHandle joint_limits_handle(joint_position_handle, limits, soft_limits_[i]);
+      position_joint_soft_limits_interface_.registerHandle(joint_limits_handle);
     }
     else if (joint.getActuationMode() == march::ActuationMode::torque)
     {
       // Create effort joint interface
-      JointHandle jointEffortHandle(jointStateHandle, &joint_effort_command_[i]);
-      effort_joint_interface_.registerHandle(jointEffortHandle);
+      JointHandle joint_effort_handle_(joint_state_handle, &joint_effort_command_[i]);
+      effort_joint_interface_.registerHandle(joint_effort_handle_);
 
       // Create joint effort limit interface
-      EffortJointSoftLimitsHandle jointEffortLimitsHandle(jointEffortHandle, limits, soft_limits_[i]);
-      effortJointSoftLimitsInterface.registerHandle(jointEffortLimitsHandle);
+      EffortJointSoftLimitsHandle joint_effort_limits_handle(joint_effort_handle_, limits, soft_limits_[i]);
+      effort_joint_soft_limits_interface_.registerHandle(joint_effort_limits_handle);
     }
 
     // Create velocity joint interface
-    JointHandle jointVelocityHandle(jointStateHandle, &joint_velocity_command_[i]);
-    velocity_joint_interface_.registerHandle(jointVelocityHandle);
+    JointHandle joint_velocity_handle(joint_state_handle, &joint_velocity_command_[i]);
+    velocity_joint_interface_.registerHandle(joint_velocity_handle);
 
     // Create march_state interface
-    MarchTemperatureSensorHandle marchTemperatureSensorHandle(joint_names_[i], &joint_temperature_[i],
-                                                              &joint_temperature_variance_[i]);
-    march_temperature_interface.registerHandle(marchTemperatureSensorHandle);
+    MarchTemperatureSensorHandle temperature_sensor_handle(joint_names_[i], &joint_temperature_[i],
+                                                           &joint_temperature_variance_[i]);
+    march_temperature_interface_.registerHandle(temperature_sensor_handle);
 
     // Enable high voltage on the IMC
     if (joint.canActuate())
     {
-      if (hasPowerDistributionBoard)
+      if (has_power_distribution_board_)
       {
         int net_number = joint.getNetNumber();
         if (net_number != -1)
         {
-          marchRobot.getPowerDistributionBoard()->getHighVoltage().setNetOnOff(true, net_number);
+          march_robot_.getPowerDistributionBoard()->getHighVoltage().setNetOnOff(true, net_number);
         }
         else
         {
           ROS_FATAL("Joint %s has no high voltage net number", joint.getName().c_str());
-          throw std::runtime_error("Joint has no high voltage net number");
+          return false;
         }
       }
       joint.prepareActuation();
@@ -201,33 +211,25 @@ void MarchHardwareInterface::init()
     }
   }
   ROS_INFO("Successfully actuated all joints");
-  controller_manager_.reset(new controller_manager::ControllerManager(this, nh_));
-}
-
-void MarchHardwareInterface::update(const ros::Duration& elapsed_time)
-{
-  read(elapsed_time);
-  validate();
-  controller_manager_->update(ros::Time::now(), elapsed_time);
-  write(elapsed_time);
+  return true;
 }
 
 void MarchHardwareInterface::validate()
 {
-  for (int i = 0; i < num_joints_; i++)
+  for (size_t i = 0; i < num_joints_; i++)
   {
     this->outsideLimitsCheck(i);
     this->iMotionCubeStateCheck(i);
   }
 }
 
-void MarchHardwareInterface::read(const ros::Duration& elapsed_time)
+void MarchHardwareInterface::read(const ros::Time& /* time */, const ros::Duration& elapsed_time)
 {
-  for (int i = 0; i < num_joints_; i++)
+  for (size_t i = 0; i < num_joints_; i++)
   {
-    float oldPosition = joint_position_[i];
+    float old_position = joint_position_[i];
 
-    march::Joint joint = marchRobot.getJoint(joint_names_[i]);
+    march::Joint joint = march_robot_.getJoint(joint_names_[i]);
 
     joint_position_[i] = joint.getAngleRad();
 
@@ -237,7 +239,7 @@ void MarchHardwareInterface::read(const ros::Duration& elapsed_time)
     }
 
     // Get velocity from encoder position
-    float joint_velocity = (joint_position_[i] - oldPosition) * 1 / elapsed_time.toSec();
+    float joint_velocity = (joint_position_[i] - old_position) * 1 / elapsed_time.toSec();
 
     // Apply exponential smoothing to velocity obtained from encoder with
     // alpha=0.2
@@ -250,9 +252,9 @@ void MarchHardwareInterface::read(const ros::Duration& elapsed_time)
 
   this->updateIMotionCubeState();
 
-  if (hasPowerDistributionBoard)
+  if (has_power_distribution_board_)
   {
-    power_distribution_board_read_ = *marchRobot.getPowerDistributionBoard();
+    power_distribution_board_read_ = *march_robot_.getPowerDistributionBoard();
 
     if (!power_distribution_board_read_.getHighVoltage().getHighVoltageEnabled())
     {
@@ -261,23 +263,23 @@ void MarchHardwareInterface::read(const ros::Duration& elapsed_time)
   }
 }
 
-void MarchHardwareInterface::write(const ros::Duration& elapsed_time)
+void MarchHardwareInterface::write(const ros::Time& /* time */, const ros::Duration& elapsed_time)
 {
-  for (int i = 0; i < num_joints_; i++)
+  for (size_t i = 0; i < num_joints_; i++)
   {
     // Enlarge joint_effort_command because ROS control limits the pid values to a certain maximum
     joint_effort_command_[i] = joint_effort_command_[i] * 1000;
   }
 
   // Enforce limits on all joints in effort mode
-  effortJointSoftLimitsInterface.enforceLimits(elapsed_time);
+  effort_joint_soft_limits_interface_.enforceLimits(elapsed_time);
 
   // Enforce limits on all joints in position mode
-  positionJointSoftLimitsInterface.enforceLimits(elapsed_time);
+  position_joint_soft_limits_interface_.enforceLimits(elapsed_time);
 
-  for (int i = 0; i < num_joints_; i++)
+  for (size_t i = 0; i < num_joints_; i++)
   {
-    march::Joint joint = marchRobot.getJoint(joint_names_[i]);
+    march::Joint joint = march_robot_.getJoint(joint_names_[i]);
 
     if (joint.canActuate())
     {
@@ -299,7 +301,7 @@ void MarchHardwareInterface::write(const ros::Duration& elapsed_time)
 
   this->updateAfterLimitJointCommand();
 
-  if (hasPowerDistributionBoard)
+  if (has_power_distribution_board_)
   {
     updatePowerDistributionBoard();
   }
@@ -310,19 +312,19 @@ void MarchHardwareInterface::initiateIMC()
   ROS_INFO("Resetting all IMC on initialization");
   for (const std::string& joint_name : joint_names_)
   {
-    Joint joint = marchRobot.getJoint(joint_name);
+    Joint joint = march_robot_.getJoint(joint_name);
     joint.resetIMotionCube();
   }
 
   ROS_INFO("Restarting EtherCAT");
-  marchRobot.stopEtherCAT();
-  marchRobot.startEtherCAT();
+  march_robot_.stopEtherCAT();
+  march_robot_.startEtherCAT();
 }
 
 void MarchHardwareInterface::updatePowerDistributionBoard()
 {
-  marchRobot.getPowerDistributionBoard()->setMasterOnline();
-  marchRobot.getPowerDistributionBoard()->setMasterShutDownAllowed(master_shutdown_allowed_command);
+  march_robot_.getPowerDistributionBoard()->setMasterOnline();
+  march_robot_.getPowerDistributionBoard()->setMasterShutDownAllowed(master_shutdown_allowed_command_);
   updateHighVoltageEnable();
   updatePowerNet();
 }
@@ -331,11 +333,12 @@ void MarchHardwareInterface::updateHighVoltageEnable()
 {
   try
   {
-    if (marchRobot.getPowerDistributionBoard()->getHighVoltage().getHighVoltageEnabled() != enable_high_voltage_command)
+    if (march_robot_.getPowerDistributionBoard()->getHighVoltage().getHighVoltageEnabled() !=
+        enable_high_voltage_command_)
     {
-      marchRobot.getPowerDistributionBoard()->getHighVoltage().enableDisableHighVoltage(enable_high_voltage_command);
+      march_robot_.getPowerDistributionBoard()->getHighVoltage().enableDisableHighVoltage(enable_high_voltage_command_);
     }
-    else if (!marchRobot.getPowerDistributionBoard()->getHighVoltage().getHighVoltageEnabled())
+    else if (!march_robot_.getPowerDistributionBoard()->getHighVoltage().getHighVoltageEnabled())
     {
       ROS_WARN_THROTTLE(2, "High voltage disabled");
     }
@@ -345,7 +348,7 @@ void MarchHardwareInterface::updateHighVoltageEnable()
     ROS_ERROR("%s", exception.what());
     ROS_DEBUG("Reverting the enable_high_voltage_command input, in attempt to prevent this exception is thrown "
               "again");
-    enable_high_voltage_command = !enable_high_voltage_command;
+    enable_high_voltage_command_ = !enable_high_voltage_command_;
   }
 }
 
@@ -355,11 +358,11 @@ void MarchHardwareInterface::updatePowerNet()
   {
     try
     {
-      if (marchRobot.getPowerDistributionBoard()->getHighVoltage().getNetOperational(
+      if (march_robot_.getPowerDistributionBoard()->getHighVoltage().getNetOperational(
               power_net_on_off_command_.getNetNumber()) != power_net_on_off_command_.isOnOrOff())
       {
-        marchRobot.getPowerDistributionBoard()->getHighVoltage().setNetOnOff(power_net_on_off_command_.isOnOrOff(),
-                                                                             power_net_on_off_command_.getNetNumber());
+        march_robot_.getPowerDistributionBoard()->getHighVoltage().setNetOnOff(
+            power_net_on_off_command_.isOnOrOff(), power_net_on_off_command_.getNetNumber());
       }
     }
     catch (std::exception& exception)
@@ -373,11 +376,11 @@ void MarchHardwareInterface::updatePowerNet()
   {
     try
     {
-      if (marchRobot.getPowerDistributionBoard()->getLowVoltage().getNetOperational(
+      if (march_robot_.getPowerDistributionBoard()->getLowVoltage().getNetOperational(
               power_net_on_off_command_.getNetNumber()) != power_net_on_off_command_.isOnOrOff())
       {
-        marchRobot.getPowerDistributionBoard()->getLowVoltage().setNetOnOff(power_net_on_off_command_.isOnOrOff(),
-                                                                            power_net_on_off_command_.getNetNumber());
+        march_robot_.getPowerDistributionBoard()->getLowVoltage().setNetOnOff(power_net_on_off_command_.isOnOrOff(),
+                                                                              power_net_on_off_command_.getNetNumber());
       }
     }
     catch (std::exception& exception)
@@ -396,19 +399,14 @@ void MarchHardwareInterface::updateAfterLimitJointCommand()
     return;
   }
 
-  // Clear msg of AfterLimitJointCommand
-  after_limit_joint_command_pub_->msg_.name.clear();
-  after_limit_joint_command_pub_->msg_.position_command.clear();
-  after_limit_joint_command_pub_->msg_.effort_command.clear();
-
-  for (int i = 0; i < num_joints_; i++)
+  after_limit_joint_command_pub_->msg_.header.stamp = ros::Time::now();
+  for (size_t i = 0; i < num_joints_; i++)
   {
-    march::Joint joint = marchRobot.getJoint(joint_names_[i]);
+    march::Joint joint = march_robot_.getJoint(joint_names_[i]);
 
-    after_limit_joint_command_pub_->msg_.header.stamp = ros::Time::now();
-    after_limit_joint_command_pub_->msg_.name.push_back(joint.getName());
-    after_limit_joint_command_pub_->msg_.position_command.push_back(joint_position_command_[i]);
-    after_limit_joint_command_pub_->msg_.effort_command.push_back(joint_effort_command_[i]);
+    after_limit_joint_command_pub_->msg_.name[i] = joint.getName();
+    after_limit_joint_command_pub_->msg_.position_command[i] = joint_position_command_[i];
+    after_limit_joint_command_pub_->msg_.effort_command[i] = joint_effort_command_[i];
   }
 
   after_limit_joint_command_pub_->unlockAndPublish();
@@ -420,58 +418,46 @@ void MarchHardwareInterface::updateIMotionCubeState()
   {
     return;
   }
-  // Clear msg of IMotionCubeStates
-  imc_state_pub_->msg_.joint_names.clear();
-  imc_state_pub_->msg_.status_word.clear();
-  imc_state_pub_->msg_.detailed_error.clear();
-  imc_state_pub_->msg_.motion_error.clear();
-  imc_state_pub_->msg_.state.clear();
-  imc_state_pub_->msg_.detailed_error_description.clear();
-  imc_state_pub_->msg_.motion_error_description.clear();
-  imc_state_pub_->msg_.motor_current.clear();
-  imc_state_pub_->msg_.motor_voltage.clear();
 
-  for (int i = 0; i < num_joints_; i++)
+  imc_state_pub_->msg_.header.stamp = ros::Time::now();
+  for (size_t i = 0; i < num_joints_; i++)
   {
-    march::IMotionCubeState iMotionCubeState = marchRobot.getJoint(joint_names_[i]).getIMotionCubeState();
-    imc_state_pub_->msg_.header.stamp = ros::Time::now();
-    imc_state_pub_->msg_.joint_names.push_back(joint_names_[i]);
-    imc_state_pub_->msg_.status_word.push_back(iMotionCubeState.statusWord);
-    imc_state_pub_->msg_.detailed_error.push_back(iMotionCubeState.detailedError);
-    imc_state_pub_->msg_.motion_error.push_back(iMotionCubeState.motionError);
-    imc_state_pub_->msg_.state.push_back(iMotionCubeState.state.getString());
-    imc_state_pub_->msg_.detailed_error_description.push_back(iMotionCubeState.detailedErrorDescription);
-    imc_state_pub_->msg_.motion_error_description.push_back(iMotionCubeState.motionErrorDescription);
-    imc_state_pub_->msg_.motor_current.push_back(iMotionCubeState.motorCurrent);
-    imc_state_pub_->msg_.motor_voltage.push_back(iMotionCubeState.motorVoltage);
+    march::IMotionCubeState imc_state = march_robot_.getJoint(joint_names_[i]).getIMotionCubeState();
+    imc_state_pub_->msg_.joint_names[i] = joint_names_[i];
+    imc_state_pub_->msg_.status_word[i] = imc_state.statusWord;
+    imc_state_pub_->msg_.detailed_error[i] = imc_state.detailedError;
+    imc_state_pub_->msg_.motion_error[i] = imc_state.motionError;
+    imc_state_pub_->msg_.state[i] = imc_state.state.getString();
+    imc_state_pub_->msg_.detailed_error_description[i] = imc_state.detailedErrorDescription;
+    imc_state_pub_->msg_.motion_error_description[i] = imc_state.motionErrorDescription;
+    imc_state_pub_->msg_.motor_current[i] = imc_state.motorCurrent;
+    imc_state_pub_->msg_.motor_voltage[i] = imc_state.motorVoltage;
   }
 
   imc_state_pub_->unlockAndPublish();
 }
 
-void MarchHardwareInterface::iMotionCubeStateCheck(int joint_index)
+void MarchHardwareInterface::iMotionCubeStateCheck(size_t joint_index)
 {
+  march::IMotionCubeState imc_state = march_robot_.getJoint(joint_names_[joint_index]).getIMotionCubeState();
+  if (imc_state.state == march::IMCState::FAULT)
   {
-    march::IMotionCubeState iMotionCubeState = marchRobot.getJoint(joint_names_[joint_index]).getIMotionCubeState();
-    if (iMotionCubeState.state == march::IMCState::FAULT)
-    {
-      std::ostringstream errorStream;
-      errorStream << "IMotionCube of joint " << joint_names_[joint_index].c_str() << " is in fault state "
-                  << iMotionCubeState.state.getString() << std::endl;
-      errorStream << "Detailed Error: " << iMotionCubeState.detailedErrorDescription << "("
-                  << iMotionCubeState.detailedError << ")" << std::endl;
-      errorStream << "Motion Error: " << iMotionCubeState.motionErrorDescription << "(" << iMotionCubeState.motionError
-                  << ")" << std::endl;
+    std::ostringstream error_stream;
+    error_stream << "IMotionCube of joint " << joint_names_[joint_index].c_str() << " is in fault state "
+                 << imc_state.state.getString() << std::endl;
+    error_stream << "Detailed Error: " << imc_state.detailedErrorDescription << "(" << imc_state.detailedError << ")"
+                 << std::endl;
+    error_stream << "Motion Error: " << imc_state.motionErrorDescription << "(" << imc_state.motionError << ")"
+                 << std::endl;
 
-      ROS_FATAL("%s", errorStream.str().c_str());
-      throw std::runtime_error(errorStream.str());
-    }
+    ROS_FATAL("%s", error_stream.str().c_str());
+    throw std::runtime_error(error_stream.str());
   }
 }
 
-void MarchHardwareInterface::outsideLimitsCheck(int joint_index)
+void MarchHardwareInterface::outsideLimitsCheck(size_t joint_index)
 {
-  march::Joint joint = marchRobot.getJoint(joint_names_[joint_index]);
+  march::Joint joint = march_robot_.getJoint(joint_names_[joint_index]);
   if (joint_position_[joint_index] < soft_limits_[joint_index].min_position ||
       joint_position_[joint_index] > soft_limits_[joint_index].max_position)
   {
@@ -481,12 +467,12 @@ void MarchHardwareInterface::outsideLimitsCheck(int joint_index)
 
     if (joint.canActuate())
     {
-      std::ostringstream errorStream;
-      errorStream << "Joint " << joint_names_[joint_index].c_str() << " is out of its soft limits ("
-                  << soft_limits_[joint_index].min_position << ", " << soft_limits_[joint_index].max_position
-                  << "). Actual position: " << joint_position_[joint_index];
-      ROS_FATAL("%s", errorStream.str().c_str());
-      throw ::std::runtime_error(errorStream.str());
+      std::ostringstream error_stream;
+      error_stream << "Joint " << joint_names_[joint_index].c_str() << " is out of its soft limits ("
+                   << soft_limits_[joint_index].min_position << ", " << soft_limits_[joint_index].max_position
+                   << "). Actual position: " << joint_position_[joint_index];
+      ROS_FATAL("%s", error_stream.str().c_str());
+      throw ::std::runtime_error(error_stream.str());
     }
   }
 }
